Renamed impl files to private and updated code style document (#2303)
* changed _impl files to _private

* renamed _impl files to _private

* renamed _impl files to_private

* updated code style doc for _impl to _private file changes

* deleted extra Software folder and updated BUILD with the new file names

* updated #include _impl files to _private

* updated #include files from _impl to _private
