+++ src/firmware/app/control/trajectory_planner.c
@@ -4,7 +4,7 @@
 #include <math.h>
 
 #include "assert.h"
-#include "firmware/app/control/trajectory_planner_impl.h"
+#include "firmware/app/control/trajectory_planner_private.h"
 #include "firmware/shared/math/polynomial_1d.h"
 #include "firmware/shared/math/polynomial_2d.h"
 #include "firmware/shared/math/tbots_math.h"
