+++ src/firmware/app/control/trajectory_planner_private.h
