+++ src/firmware/app/control/trajectory_planner_test.cpp
@@ -2,7 +2,7 @@ extern "C"
 {
 #include "firmware/app/control/trajectory_planner.h"
 
-#include "firmware/app/control/trajectory_planner_impl.h"
+#include "firmware/app/control/trajectory_planner_private.h"
 }
 
 #include <gtest/gtest.h>
